+++ tastee-backend/src/main/java/com/tastee/tastee_backend/database/UserRepo.java
@@ -1,13 +1,13 @@
 package com.tastee.tastee_backend.database;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+
 import org.springframework.stereotype.Repository;
 
 import com.tastee.tastee_backend.beans.Users;
 
 @Repository
-public interface UserRepo extends JpaRepository<Users, Integer> {
+public interface UserRepo extends JpaRepository<Users, Long> {
 
     Users findByUsername(String username);
